+++ matrix.cpp
@@ -1,14 +1,14 @@
 #include "matrix.h"
 
-Matrix::Matrix() : mat4(new float* [4])
+Matrix::Matrix() : mat4(new double* [4])
 {
 	for (int i = 0; i < 4; ++i)
 	{
-		mat4[i] = new float[4];
+		mat4[i] = new double[4];
 	}
 }
 
-Matrix::Matrix(float diagonal) : Matrix()
+Matrix::Matrix(double diagonal) : Matrix()
 {
 	mat4[0][0] = diagonal;
 	mat4[1][1] = diagonal;
@@ -27,10 +27,12 @@ Matrix::~Matrix()
 
 void Matrix::InitM(std::vector<Vector> vectors)
 {
+
 	mat4[0][0] = vectors[0].x / vectors[0].SizeVector();
 	mat4[1][0] = vectors[0].y / vectors[0].SizeVector();
 	mat4[2][0] = vectors[0].z / vectors[0].SizeVector();
 	mat4[3][0] = vectors[0].w;
+
 
 	mat4[0][1] = vectors[1].x / vectors[1].SizeVector();
 	mat4[1][1] = vectors[1].y / vectors[1].SizeVector();
@@ -41,6 +43,7 @@ void Matrix::InitM(std::vector<Vector> vectors)
 	mat4[1][2] = vectors[2].y / vectors[2].SizeVector();
 	mat4[2][2] = vectors[2].z / vectors[2].SizeVector();
 	mat4[3][2] = vectors[2].w;
+
 
 	mat4[0][3] = vectors[3].x;
 	mat4[1][3] = vectors[3].y;
